+++ package.json
@@ -9,7 +9,7 @@
     "server": "json-server data.json",
     "format": "prettier --write '{*.js,src/**/*.{js,jsx}}'",
     "lint": "eslint src/ --fix",
-    "test": "jest"
+    "test": "jest --watch"
   },
   "keywords": [
     "react",
