+++ src/__test__/components/About.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import About from '../../components/About';
 
 describe('<About />', () => {
-  const about = mount(<About />);
+  const about = shallow(<About />);
 
   test('About render', () => {
     expect(about.length).toEqual(1);
@@ -16,5 +16,4 @@ describe('<About />', () => {
   test('About haves 3 items', () => {
     expect(about.find('.About-item').length).toBeGreaterThan(2);
   });
-
 });
