+++ src/__test__/components/Header.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { mount } from 'enzyme';
+import { shallow } from 'enzyme';
 import Header from '../../components/Header';
 
 describe('<Header />', () => {
-  const header = mount(<Header />);
+  const header = shallow(<Header />);
 
   test('Header render', () => {
     expect(header.length).toEqual(1);
@@ -12,5 +12,4 @@ describe('<Header />', () => {
   test('Header title', () => {
     expect(header.find('.Header-title').length).toEqual(1);
   });
-
 });
