+++ src/components/About.js
@@ -1,9 +1,16 @@
-import React from 'react'
+import React from 'react';
 
 const About = () => {
   return (
-    <div>About</div>
-  )
-}
+    <>
+      <h2 className="About-title">About</h2>
+      <ul>
+        <li className="About-item">home</li>
+        <li className="About-item">about us</li>
+        <li className="About-item">contact</li>
+      </ul>
+    </>
+  );
+};
 
-export default About
+export default About;
