+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Header = ({ children }) => {
+  return (
+    <>
+      <h1 className="Header-title">Header</h1>
+      {children}
+    </>
+  );
+};
+
+export default Header;
