--- src/components/Header.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Header = () => {
-  return (
-    <div>Header</div>
-  )
-}
-
-export default Header
