+++ src/components/Profile.js
@@ -1,9 +1,17 @@
-import React from 'react'
+import React from 'react';
 
 const Profile = () => {
   return (
-    <div>Profile</div>
-  )
-}
+    <>
+      <h2 className="Profile-title">Profile</h2>
+      <p className="Profile-desc">
+        Lorem ipsum, dolor sit amet consectetur adipisicing elit. Possimus earum
+        veritatis nostrum optio vel itaque perspiciatis, placeat ducimus
+        consequatur delectus, illo ratione neque culpa animi aliquam, nulla
+        eveniet accusantium. In?
+      </p>
+    </>
+  );
+};
 
-export default Profile
+export default Profile;
